fix: fix status
