+++ cmd/oras/internal/display/status/progress/status_test.go
@@ -29,22 +29,20 @@ import (
 
 func Test_status_String(t *testing.T) {
 	// zero status and progress
-	s := newStatus()
+	s := newStatus(ocispec.Descriptor{
+		MediaType: "application/vnd.oci.empty.oras.test.v1+json",
+		Size:      2,
+		Digest:    "sha256:44136fa355b3678a1146ad16f7e8649e94fb4fc21fe77e8310c060f61caaff8a",
+	})
 	if status, digest := s.String(console.MinWidth); status != zeroStatus || digest != zeroDigest {
 		t.Errorf("status.String() = %v, %v, want %v, %v", status, digest, zeroStatus, zeroDigest)
 	}
 
 	// not done
 	s.update(&status{
-		prompt: "test",
-		descriptor: ocispec.Descriptor{
-			MediaType: "application/vnd.oci.empty.oras.test.v1+json",
-			Size:      2,
-			Digest:    "sha256:44136fa355b3678a1146ad16f7e8649e94fb4fc21fe77e8310c060f61caaff8a",
-		},
+		prompt:    "test",
 		startTime: time.Now().Add(-time.Minute),
 		offset:    0,
-		total:     humanize.ToBytes(2),
 	})
 	// full name
 	statusStr, digestStr := s.String(120)
@@ -70,22 +68,20 @@ func Test_status_String(t *testing.T) {
 
 func Test_status_String_zeroWidth(t *testing.T) {
 	// zero status and progress
-	s := newStatus()
+	s := newStatus(ocispec.Descriptor{
+		MediaType: "application/vnd.oci.empty.oras.test.v1+json",
+		Size:      0,
+		Digest:    "sha256:e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
+	})
 	if status, digest := s.String(console.MinWidth); status != zeroStatus || digest != zeroDigest {
 		t.Errorf("status.String() = %v, %v, want %v, %v", status, digest, zeroStatus, zeroDigest)
 	}
 
 	// not done
 	s.update(&status{
-		prompt: "test",
-		descriptor: ocispec.Descriptor{
-			MediaType: "application/vnd.oci.empty.oras.test.v1+json",
-			Size:      0,
-			Digest:    "sha256:e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
-		},
+		prompt:    "test",
 		startTime: time.Now().Add(-time.Minute),
 		offset:    0,
-		total:     humanize.ToBytes(0),
 	})
 	// not done
 	statusStr, digestStr := s.String(120)
@@ -105,7 +101,7 @@ func Test_status_String_zeroWidth(t *testing.T) {
 }
 func Test_status_durationString(t *testing.T) {
 	// zero duration
-	s := newStatus()
+	s := newStatus(ocispec.Descriptor{})
 	if d := s.durationString(); d != zeroDuration {
 		t.Errorf("status.durationString() = %v, want %v", d, zeroDuration)
 	}
